Move lidarr base entity to separate module
